+++ src/Phamda.php
@@ -34,6 +34,8 @@ class Phamda
      * ```
      *
      * @return Placeholder
+     *
+     * @deprecated Since version 0.7, to be removed in 0.8. Use `flip` or custom closures instead.
      */
     public static function _()
     {